Add offset and create hook useCloseOnOverlay
